+++ src/fields/GInput.tsx
@@ -1,23 +1,33 @@
-import React, {forwardRef, ReactNode, useEffect, useMemo} from 'react';
+import React, {forwardRef, memo, type ReactNode, useEffect, useMemo} from 'react';
 
-import { _debounce } from '../helpers';
-import type { GInputProps, GInputState, GElementProps } from '.';
-import {useFormSelector, createSelector, useFormStore} from "../form-context";
+import {_debounce} from '../helpers';
+import type {GInputProps, GInputState, GElementProps} from '.';
+import {useFormSelector, useFormStore} from "../form-context";
+import {makeSelectFields} from "../selectors";
 
-const selectFields = [(state) => state.fields];
-
-export const makeSelectFields = (keys: string[]) =>
-    createSelector(
-        selectFields,
-        (fields) => {
-            const selected = keys.map((key) => fields[key]).filter(Boolean);
-            return selected.length ? selected : null;
-        }
-    );
-
-export const GInput = forwardRef<HTMLInputElement, GInputProps>(({ formKey, element, title, type, validatorKey, fetch, fetchDeps = ['unknown'], optimized, defaultChecked, defaultValue, checked, value, debounce=300, ...rest }, ref) => {
+const _GInput = forwardRef<HTMLInputElement, GInputProps>(({
+    formKey,
+    element,
+    title,
+    type = 'text',
+    fetch,
+    fetchDeps,
+    optimized,
+    debounce = 300,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    defaultChecked,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    defaultValue,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    checked,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    validatorKey,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    value,
+    ...rest
+}, ref) => {
     const inputState = useFormSelector(state => state.fields[formKey]);
-    const handlers = useFormStore().handlers;
+    const store = useFormStore();
 
     const _element = useMemo(() => {
         let value: any, checked;
@@ -37,20 +47,32 @@ export const GInput = forwardRef<HTMLInputElement, GInputProps>(({ formKey, elem
             title: title || inputState.errorText
         };
 
-        if (!handlers.formOptimized || !optimized) {
-            _props.onBlur = (e) => {
-                handlers._viHandler(inputState, e);
-                rest.onBlur && rest.onBlur(e);
-            };
-            _props.onInvalid = (e) => {
-                e.preventDefault(); // hide default browser validation tooltip
-                handlers._viHandler(inputState, e);
-                rest.onInvalid && rest.onInvalid(e);
-            };
-            _props.onChange = (e, unknown?: { value: unknown } | string | number) => {
-                handlers._updateInputHandler(inputState, e, unknown);
-                rest.onChange && rest.onChange(e);
-            };
+        if (!store.handlers.optimized || !optimized) {
+            _props.onBlur = rest.onBlur ?
+                (e) => {
+                    store.handlers._viHandler(inputState, e);
+                    rest.onBlur!(e);
+                } : (e) => {
+                    store.handlers._viHandler(inputState, e);
+                };
+
+            _props.onInvalid = rest.onInvalid ?
+                (e) => {
+                    e.preventDefault(); // hide default browser validation tooltip
+                    store.handlers._viHandler(inputState, e);
+                    rest.onInvalid!(e);
+                } : (e) => {
+                    e.preventDefault(); // hide default browser validation tooltip
+                    store.handlers._viHandler(inputState, e);
+                };
+
+            _props.onChange = rest.onChange ?
+                (e, unknown?: { value: unknown } | string | number) => {
+                    store.handlers._updateInputHandler(inputState, e, unknown);
+                    rest.onChange!(e);
+                } : (e, unknown?: { value: unknown } | string | number) => {
+                    store.handlers._updateInputHandler(inputState, e, unknown);
+                };
         }
 
         if (element) {
@@ -62,22 +84,20 @@ export const GInput = forwardRef<HTMLInputElement, GInputProps>(({ formKey, elem
         );
     }, [inputState, element]);
 
-    // const selector = useMemo(() => makeSelectFields(fetchDeps), [fetchDeps]);
     const _fetchDeps = useFormSelector(makeSelectFields(fetchDeps));
-
     const stableFetchDeps = useMemo(() => JSON.stringify(_fetchDeps), [_fetchDeps]);
 
     useEffect(() => {
         if (fetch) {
             inputState.dispatchChanges = (changes: Partial<GInputState>) =>
-                handlers._dispatchChanges(changes, formKey);
+                store.handlers._dispatchChanges(changes, formKey);
 
             _debounce(debounce, `${inputState.gid}-fetch`).then(() => {
-                const res = fetch(inputState, _fetchDeps);
+                const res = fetch(inputState, store.getState().fields);
                 if (res instanceof Promise) {
-                    res.then((state) => state && handlers._dispatchChanges(state, formKey));
-                } else {
-                    res && handlers._dispatchChanges(res, formKey);
+                    res.then((state) => state && store.handlers._dispatchChanges(state, formKey));
+                } else if (res) {
+                    store.handlers._dispatchChanges(res, formKey);
                 }
             });
         }
@@ -85,3 +105,5 @@ export const GInput = forwardRef<HTMLInputElement, GInputProps>(({ formKey, elem
 
     return _element;
 });
+
+export const GInput = memo(_GInput);